requirements.txt
fastapi
uvicorn
torch
streamlit
pandas
scikit-learn
mysql-connector-python
SQLAlchemy
pymysql
python-dotenv
requests